docs: Fix import error.
```
Running Sphinx v1.8.5
loading translations [en]... done

Traceback (most recent call last):
  File "/home/docs/checkouts/readthedocs.org/user_builds/mithro-vtr/envs/latest/lib/python3.7/site-packages/sphinx/registry.py", line 472, in load_extension
    mod = __import__(extname, None, None, ['setup'])
  File "/home/docs/checkouts/readthedocs.org/user_builds/mithro-vtr/checkouts/latest/doc/_exts/sdcdomain/__init__.py", line 1, in <module>
    from sdcdomain import setup
ImportError: cannot import name 'setup' from 'sdcdomain' (/home/docs/checkouts/readthedocs.org/user_builds/mithro-vtr/checkouts/latest/doc/_exts/sdcdomain/__init__.py)

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/home/docs/checkouts/readthedocs.org/user_builds/mithro-vtr/envs/latest/lib/python3.7/site-packages/sphinx/cmd/build.py", line 303, in build_main
    args.tags, args.verbosity, args.jobs, args.keep_going)
  File "/home/docs/checkouts/readthedocs.org/user_builds/mithro-vtr/envs/latest/lib/python3.7/site-packages/sphinx/application.py", line 228, in __init__
    self.setup_extension(extension)
  File "/home/docs/checkouts/readthedocs.org/user_builds/mithro-vtr/envs/latest/lib/python3.7/site-packages/sphinx/application.py", line 449, in setup_extension
    self.registry.load_extension(self, extname)
  File "/home/docs/checkouts/readthedocs.org/user_builds/mithro-vtr/envs/latest/lib/python3.7/site-packages/sphinx/registry.py", line 475, in load_extension
    raise ExtensionError(__('Could not import extension %s') % extname, err)
sphinx.errors.ExtensionError: Could not import extension sdcdomain (exception: cannot import name 'setup' from 'sdcdomain' (/home/docs/checkouts/readthedocs.org/user_builds/mithro-vtr/checkouts/latest/doc/_exts/sdcdomain/__init__.py))

Extension error:
Could not import extension sdcdomain (exception: cannot import name 'setup' from 'sdcdomain' (/home/docs/checkouts/readthedocs.org/user_builds/mithro-vtr/checkouts/latest/doc/_exts/sdcdomain/__init__.py))
```
